Add title attribute to paper hyperlinks
This adds two new generated meta-data files to the repo, index.json and
paper_titles.txt. The former is downloaded from wg21.link and then
processed by Python to produce the latter. But if Python is not
installed an empty paper_titles.txt file will be created, and no title
attributes added to the generated HTML.