Support simple markdown-like syntax for inline code
This adds a preprocessing step to the issue text before starting to
process it as XML. Any block delimited by triple-backticks (alone on a
line) will be enclosed in `<pre>` and `<code>` elements, Any text
enclosed by a single backtick (without any intervening newlines) will be
enclosed in a `<code>` element.

In both cases, all occurrences of the characters `&` and `<` and `>`
inside the code element will be replaced with the corresponding HTML
entities.

This makes it much simpler to write (and copy & paste) C++ code into
issues.