Add (i) snapshot links to output of <iref> elements
The links generated by <iref> always go to the full issues list, which
can be slow to load. In those full lists the issue number includes a (i)
superscript that links to the single-page view of the issue. This change
makes <iref> elements also add a (i) superscript linking to the
single-page version of the issue.