Simplify workflow for deploying new HTML pages
Cloning the entire repo is very slow, due to the gh-pages branch
containing nearly 2GiB of static html files. The current version of the
deploy.yml workflow uses actions/cache@v4 to cache a copy of the full
repo which can be retrieved much faster than doing a clone. This makes
the workflow faster, but is complex to understand. It also doesn't help
if no changes have been pushed to the master branch for more than a
week, because the cache expires and so a full clone is needed anyway.

This change attempts to remove the cache so that the workflow is simpler
to understand and more robust.

To avoid cloning the full repo, the initial clone uses the magic number
fetch-depth:2147483647 i.e. INT_MAX. According to the documentation at
https://git-scm.com/docs/shallow this means infinite depth. However,
unlike fetch-depth:0 it implies the --single-branch option for the
clone, so only the master branch is fetched and not the huge gh-pages
branch.

The gh-pages branch is still needed in order to push to that branch, but
we clone that in a separate step (into a sub-directory) using
fetch-depth:1 to avoid fetching the full history. That step doesn't need
to change, as it wasn't using the cache anyway.