Prevent concurrent deployments from the deploy.yml workflow
This avoids running the workflow for every push, so that we don't do
redundant work that will be replaced immediately by a later deployment.

If an existing deploy workflow is running, pushing new commits will not
run the deploy workflow for the new commits until the first workflow
finishes.  If more than one new push happens while the workflow is
running, the workflow will only be run for the last push.

Also add 'workflow_dispatch:' to allow running the workflow manually.