Streamline deploy.yml workflow to do less work, but in parallel
Remove the separate steps that run 'make dates' and 'make pgms'.

Only the bin/lists executable is needed to generate the HTML lists, and
that will be built as a prerequisite of the 'lists' target anyway. We
can remove the separate step that runs 'make pgms' to build all the
executables, saving a few seconds by not compiling+linking sources we
don't need for this workflow.

We can also remove the separate step that runs 'make dates' because
that's also a prerequisite of the 'lists' target.

After this change we only run 'make lists' but we do that with the -j
option so that fetching and generating the meta-data can be done in
parallel. Previously only compiling the sources was done in parallel.