Make most lwg::report_generator member functions const
The remaining functions access the section_db and so can't be const yet.
Ideally we would avoid making any changes to that section_db after
parsing the XML files and converting the issue text to HTML. That would
allow all functions that generate HTML files to be const, which would
allow them to run in parallel. Generating the individual HTML files
currently takes about 25% of the total time to generate the lists, but
in theory in should be possible to generate those files concurrently.

Also make the paper_title_attr function take a const reference to the
lwg::metadata object. It needs to lookup paper titles in the map, but it
can use map::find instead of map::operator[], so that it doesn't need a
non-const reference.