Generate issue lists concurrently
By using four threads to create the issues lists we reduce the total
runtime from 1.5s to 1s. We could do much better if we could generate
the individual HTML files for each issue in parallel, or even just do
that in parallel with creating the main lists. Currently we have to
finish generating the main lists before starting to generate the
individual issues.